requirements.txt
torch>=1.13
transformers>=4.30
sentence-transformers>=2.2.2
faiss-cpu>=1.7.3
pandas
scikit-learn
streamlit
gradio
tqdm
rapidfuzz            
